refactor of iceberg handler
