+++ integration/spark/spark3/src/main/java/io/openlineage/spark3/agent/lifecycle/plan/catalog/iceberg/BigQueryMetastoreCatalogTypeHandler.java
@@ -32,11 +32,14 @@ class BigQueryMetastoreCatalogTypeHandler extends BaseCatalogTypeHandler {
   }
 
   @Override
-  Optional<DatasetIdentifier> getIdentifier(
+  Optional<DatasetIdentifier.Symlink> getSymlinkIdentifiers(
       SparkSession session, Map<String, String> catalogConf, String table) {
     String warehouseLocation = catalogConf.get(CatalogProperties.WAREHOUSE_LOCATION);
+    DatasetIdentifier di =
+        FilesystemDatasetUtils.fromLocationAndName(new Path(warehouseLocation).toUri(), table);
     return Optional.of(
-        FilesystemDatasetUtils.fromLocationAndName(new Path(warehouseLocation).toUri(), table));
+        new DatasetIdentifier.Symlink(
+            di.getName(), di.getNamespace(), DatasetIdentifier.SymlinkType.TABLE));
   }
 
   @Override
