+++ integration/spark/spark3/src/main/java/io/openlineage/spark3/agent/lifecycle/plan/catalog/iceberg/RestCatalogTypeHandler.java
@@ -9,20 +9,25 @@ import static io.openlineage.spark3.agent.lifecycle.plan.catalog.iceberg.Iceberg
 import static io.openlineage.spark3.agent.lifecycle.plan.catalog.iceberg.IcebergHandler.TYPE;
 
 import io.openlineage.client.utils.DatasetIdentifier;
+import io.openlineage.spark.agent.util.PathUtils;
+import io.openlineage.spark3.agent.lifecycle.plan.catalog.MissingDatasetIdentifierCatalogException;
 import java.net.URI;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.hadoop.fs.Path;
 import org.apache.iceberg.CatalogProperties;
 import org.apache.spark.sql.SparkSession;
+import org.apache.spark.sql.connector.catalog.Identifier;
+import org.apache.spark.sql.connector.catalog.TableCatalog;
 
 @Slf4j
 class RestCatalogTypeHandler extends BaseCatalogTypeHandler {
 
-  private static final String REST_CATALOG_TYPE = "rest";
-  private static final String BIGLAKE_CATALOG_URI = "https://biglake.googleapis.com/";
+  protected static final String REST_CATALOG_TYPE = "rest";
+  protected static final String BIGLAKE_CATALOG_URI = "https://biglake.googleapis.com/";
 
   @Override
   String getType() {
@@ -31,18 +36,39 @@ class RestCatalogTypeHandler extends BaseCatalogTypeHandler {
 
   @Override
   boolean matchesCatalogType(Map<String, String> catalogConf) {
-    return "rest".equalsIgnoreCase(catalogConf.get(TYPE))
-        || catalogConf.containsKey(CATALOG_IMPL)
-            && catalogConf.get(CATALOG_IMPL).endsWith("RESTCatalog");
+    return isRestCatalog(catalogConf);
+    //        && !catalogConf.getOrDefault(CatalogProperties.URI,
+    // "").startsWith(BIGLAKE_CATALOG_URI);
+  }
+
+  @Override
+  DatasetIdentifier getPrimaryIdentifier(
+      SparkSession session,
+      Map<String, String> catalogConf,
+      Identifier identifier,
+      TableCatalog tableCatalog) {
+
+    Optional<Path> maybeTableLocation = getTableLocation(identifier, tableCatalog);
+    String warehouseLocation = catalogConf.get(CatalogProperties.WAREHOUSE_LOCATION);
+    if (!maybeTableLocation.isPresent() && warehouseLocation == null) {
+      log.debug(
+          "The catalog type is 'rest' and the table location and warehouse location is empty. This is likely a table that is being created");
+      throw new MissingDatasetIdentifierCatalogException(
+          "No table location found. Probably needs to create table first");
+    }
+    return PathUtils.fromPath(
+        maybeTableLocation.orElseGet(
+            () -> defaultTableLocation(new Path(warehouseLocation), identifier)));
   }
 
   @Override
   @SneakyThrows
-  Optional<DatasetIdentifier> getIdentifier(
+  Optional<DatasetIdentifier.Symlink> getSymlinkIdentifiers(
       SparkSession session, Map<String, String> catalogConf, String table) {
     String confUri = catalogConf.get(CatalogProperties.URI);
     String uri = new URI(confUri).toString();
-    return Optional.of(new DatasetIdentifier(table, uri));
+    return Optional.of(
+        new DatasetIdentifier.Symlink(table, uri, DatasetIdentifier.SymlinkType.TABLE));
   }
 
   @Override
@@ -53,5 +79,11 @@ class RestCatalogTypeHandler extends BaseCatalogTypeHandler {
       return properties;
     }
     return super.catalogProperties(catalogConf);
+  }
+
+  protected static boolean isRestCatalog(Map<String, String> catalogConf) {
+    return "rest".equalsIgnoreCase(catalogConf.get(TYPE))
+        || catalogConf.containsKey(CATALOG_IMPL)
+            && catalogConf.get(CATALOG_IMPL).endsWith("RESTCatalog");
   }
 }
